feat: altera tecnologia de armazenamento de informações.
Reimplementa e corrige algumas migrations para adequar ao padrão do postgresql.

+++ src/main/resources/db/migration/V5__alter-table-pacientes-add-column-ativo.sql
@@ -1,2 +1,2 @@
-alter table pacientes add ativo tinyint;
+alter table pacientes add ativo int;
 update pacientes set ativo = 1;
